moved implementation to cpp file.
added better comments for the expand method.
